ci: fix release auth (see body)
The release workflow accidently commited the authToken in the .npmrc file by changesets. I'm so sorry for this indcident. I have ensured the old token is gone and that no unauthorized versions were published. Fortunately, the token didn't have the correct permissions to publish anyway, hence the failed release workflows.